requirements.txt
numpy>=1.26,<3.0
ultralytics==8.3.88
opencv-python==4.11.0.86
PyYAML==6.0.2
